requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
PyJWT==2.8.0
Flask-Cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Werkzeug==3.0.3 